Refactor ImageBackingStore and ROMDrive to separate files.
Makes ZuluSCSI_disk.cpp easier to maintain.
No functional changes.
